% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/call_census_api.R
\name{call_api_once}
\alias{call_api_once}
\title{Call Census API for a set of variables}
\usage{
call_api_once(variables_to_get, geoid, call_start)
}
\arguments{
\item{variables_to_get}{the variable name for the Census API call,
defined at \url{http://api.census.gov/}}

\item{geoid}{A character string with a FIPS code, between 2 and 15 digits long.}
}
\value{
A code{data.frame} with the requested variables at the requested
  geography.
}
\description{
This is an internal function and is not intended for users. See instead
\link{call_census_api}.
}

