% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/censusr.R
\docType{package}
\name{censusr}
\alias{censusr}
\alias{censusr-package}
\title{censusr: A package to download Census API data}
\description{
The censusr package provides principally one main function: call_census_api
allows users to pass a list of variables and a list of geographies, and returns
a data_frame with the requested data.
}
