% Generated by roxygen2 (4.0.2.9000): do not edit by hand
% Please edit documentation in R/get_data.R
\name{get_data}
\alias{get_data}
\title{Function to get data}
\usage{
get_data(resource = "occurrence", version = "v1",
  base_url = "wabarr.com/census", ...)
}
\arguments{
\item{resource}{API resource to access. Default is occurrence.}

\item{version}{Version of the API to use.  Default is "v1"}

\item{base_url}{Url to use to access API. Useful to override to local server during development. This should probably be removed once API is production ready}

\item{...}{An arbitrary number of filter critera in the following form: genusName__contains="Aepycer"}
}
\description{
get data
}
\examples{
get_data(source)
}
\keyword{bovid}
\keyword{census}
\keyword{data}

