% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/call_census_api.R
\name{call_acs_api}
\alias{call_acs_api}
\title{Retrieve data from the Census ACS API}
\usage{
call_acs_api(variables_to_get, geoids)
}
\arguments{
\item{variables_to_get}{the variable name for the Census API call,
defined at \url{http://api.census.gov/}}

\item{geoids}{a vector of FIPS codes; must be at least to the county (5-digit)
level, and can accept down to block groups (12-digit).}
}
\value{
a data_frame with each requested variable at each requested geography.
}
\description{
Returns Census ACS data for the  ACS aggregation for requested
variables at requested geographies.
}

