% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/call_census_api.R
\name{get_geo_url}
\alias{get_geo_url}
\title{Construct a geography request string from a FIPS Code}
\usage{
get_geo_url(geoid, allgeos)
}
\arguments{
\item{geoid}{A character string with a FIPS code, between 2 and 15 digits long.}

\item{allgeos}{(optional) A string identifying the type of geography for
which to collect data within the the requested \code{geoids}. Must be one
of \code{c('co', 'tr', 'bg', 'bl')}. For instance, if \code{allgeos =
"bg"}, will return all block groups within the given \code{geoids}.}
}
\value{
A string with the FIPS formatted for an API request.
}
\description{
Construct a geography request string from a FIPS Code
}
