% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/call_census_api.R
\name{get_geo_url}
\alias{get_geo_url}
\title{Construct a geography request string from a FIPS Code}
\usage{
get_geo_url(geoid, allgeos)
}
\arguments{
\item{geoid}{A character string with a FIPS code, between 2 and 15 digits long.}

\item{allgeos}{(optional) A string identifying the type of geography to
collect data for that live within the geoids (i.e., "co", "tr",
"bg", "bl").}
}
\value{
A string with the FIPS formatted for an API request.
}
\description{
Construct a geography request string from a FIPS Code
}

